RALPH: Review - move runtime update payload helper
