+++ packages/server-core/src/sessions/runtime-config.ts
@@ -1,4 +1,4 @@
-import type { AgentProvider, LlmAuthType } from '@craft-agent/shared/agent/backend'
+import type { AgentProvider, BackendRuntimeUpdate, LlmAuthType, ResolvedBackendContext } from '@craft-agent/shared/agent/backend'
 import { isCompatProvider, modelSupportsImages, type LlmConnection } from '@craft-agent/shared/config'
 import type { FileAttachment } from '@craft-agent/shared/protocol'
 
@@ -31,6 +31,42 @@ function normalizeCustomModels(connection: LlmConnection): Array<Record<string, 
       })
     })
     .sort((a, b) => String(a.id).localeCompare(String(b.id)))
+}
+
+function buildRuntimeCustomModels(connection: LlmConnection): NonNullable<BackendRuntimeUpdate['runtime']>['customModels'] {
+  return connection.models?.map(model => {
+    if (typeof model === 'string') return model
+
+    const supportsImages = typeof model.supportsImages === 'boolean' ? model.supportsImages : undefined
+    if (model.contextWindow || supportsImages !== undefined) {
+      return {
+        id: model.id,
+        ...(model.contextWindow ? { contextWindow: model.contextWindow } : {}),
+        ...(supportsImages !== undefined ? { supportsImages } : {}),
+      }
+    }
+
+    return model.id
+  })
+}
+
+/**
+ * Build the in-place backend runtime update payload from a resolved backend
+ * context. This mirrors the runtime envelope accepted by live Pi subprocesses.
+ */
+export function buildRuntimeConfigUpdate(backendContext: ResolvedBackendContext): BackendRuntimeUpdate {
+  const connection = backendContext.connection
+  return {
+    model: backendContext.resolvedModel,
+    providerType: connection?.providerType,
+    authType: backendContext.authType,
+    runtime: connection ? {
+      baseUrl: connection.baseUrl,
+      piAuthProvider: connection.piAuthProvider,
+      customEndpoint: connection.customEndpoint,
+      customModels: buildRuntimeCustomModels(connection),
+    } : undefined,
+  }
 }
 
 /**
